Adopted overline for unrepresentable colors
One of the main challenges of oklch is that not all colors are
representable in rgb. Björn Ottosson's clipping algorithm attempts to
find the nearest rgb color, but knowing when clipping occured may be
useful for minimizing clipping and related artifacts.

An overline is a nice compromise that conveys this information without
creating problems for copy-pasting.

Internally, hex_clipoklch appends an ! to unrepresentable colors
('#e17300!', for example). The trailing ! is ignored in oklch_fromhex
and provides a nice side-channel.